+++ clinicadl/train/tasks/__init__.py
@@ -1,2 +1,1 @@
-from .available_parameters import Task
-from .base_task_config import BaseTaskConfig
+from .tasks_utils import create_training_config
